requirements.txt
# ==============================
# 데이터 수집 (Data Collection)
# ==============================
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=5.0.0

# ==============================
# 데이터 처리 (Data Processing)
# ==============================
pandas>=2.1.0
numpy>=1.24.0

# ==============================
# 컴퓨터 비전 및 OCR (Computer Vision)
# ==============================
easyocr>=1.7.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-image
albumentations
timm
einops>=0.7.0

# ==============================
# 딥러닝 및 NLP (Deep Learning & NLP)
# ==============================
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
accelerate
datasets
tokenizers
sentencepiece>=0.1.99
protobuf>=4.25.0

# ==============================
# 생성형 AI 및 번역 (Generative AI)
# ==============================
google-genai>=0.3.0
python-dotenv>=1.0.0

# ==============================
# 모델 평가 및 지표 (Evaluation Metrics)
# ==============================
bert-score>=0.3.13
sacrebleu>=2.4.0
kiwipiepy>=0.17.0

# ==============================
# 유틸리티 (Utilities)
# ==============================
tqdm>=4.66.0
pyyaml>=6.0.1
regex>=2023.10.3
